requirements.txt
python-telegram-bot==20.3
Flask==2.1.3
Flask-Babel==2.0.0
requests==2.28.2
PyPDF2==3.0.1
gunicorn==20.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.1
pytest==7.4.0

